requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.5
psycopg[binary]==3.2.3
python-dotenv==1.0.0
Werkzeug==3.0.1

